+++ app/Models/ChatroomMessage.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Model;
+namespace App\Models;
 
 use Illuminate\Database\Eloquent\Model;
 
@@ -10,6 +10,6 @@ class ChatroomMessage extends Model
 
     protected function chatroom()
     {
-        return $this->belongsTo('App\Model\Chatroom');
+        return $this->belongsTo('App\Models\Chatroom');
     }
 }
